Libretro: Read firmware files from the "system" folder rather than the "system\Firmware" folder

+++ Libretro/LibretroSoundManager.h
@@ -27,7 +27,7 @@ public:
 	{
 		if(!_skipMode && _sendAudioSample) {
 			for(uint32_t total = 0; total < sampleCount; ) {
-				total += _sendAudioSample(soundBuffer + total*2, sampleCount - total);
+				total += (uint32_t)_sendAudioSample(soundBuffer + total*2, (size_t)(sampleCount - total));
 			}
 		}
 	}
